blue/backend/requirements.txt
asgiref==3.8.1
Django==4.2.20
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
PyJWT==2.9.0
sqlparse==0.5.3
psycopg2-binary
django-prometheus==2.3.1
prometheus_client==0.21.1
mysqlclient==2.1.1